Move shared code into scripts/jest
