+++ scripts/jest/setupSpecEquivalenceReporter.js
